move tests shared across schemes to utils
