+++ src/linear_codes/multilinear_brakedown/tests.rs
@@ -8,7 +8,6 @@ mod tests {
         linear_codes::{utils::*, BrakedownPCParams, MultilinearBrakedown, PolynomialCommitment},
         LabeledPolynomial,
     };
-    use ark_bls12_377::Fq;
     use ark_bls12_377::Fr;
     use ark_bls12_381::Fr as Fr381;
     use ark_crypto_primitives::{
@@ -144,23 +143,6 @@ mod tests {
             None
         )
         .unwrap());
-    }
-
-    #[test]
-    fn test_calculate_t_with_good_parameters() {
-        assert!(calculate_t::<Fq>(128, (3, 4), 2_usize.pow(32)).unwrap() < 200);
-        assert!(calculate_t::<Fq>(256, (3, 4), 2_usize.pow(32)).unwrap() < 400);
-    }
-
-    #[test]
-    fn test_calculate_t_with_bad_parameters() {
-        calculate_t::<Fq>(
-            (Fq::MODULUS_BIT_SIZE - 60) as usize,
-            (3, 4),
-            2_usize.pow(60),
-        )
-        .unwrap_err();
-        calculate_t::<Fq>(400, (3, 4), 2_usize.pow(32)).unwrap_err();
     }
 
     fn rand_point<F: Field>(num_vars: Option<usize>, rng: &mut ChaCha20Rng) -> Vec<F> {
